+++ packages/client-ts/CHANGELOG.md
@@ -1,3 +1,6 @@
+
+- Fix: Don't ping public databases
+
 2023-02-16 (v2.1.1)
 -------------------